pyproject.toml
[project]
name = "tg-deleter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "telethon>=1.38.1",
]
